flowvars: update funcs to accept u16 id
All id's are u16, but flowvar functions would only accept u8.

Minor cleanups.

+++ src/flow-var.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2007-2010 Open Information Security Foundation
+/* Copyright (C) 2007-2013 Open Information Security Foundation
  *
  * You can copy, redistribute or modify this Program under the terms of
  * the GNU General Public License version 2 as published by the Free
@@ -37,20 +37,20 @@
 typedef struct FlowVarTypeStr {
     uint8_t *value;
     uint16_t value_len;
-}FlowVarTypeStr;
+} FlowVarTypeStr;
 
 /** Struct used to hold the integer data type for flowvars */
 typedef struct FlowVarTypeInt_ {
     uint32_t value;
-}FlowVarTypeInt;
+} FlowVarTypeInt;
 
 /** Generic Flowvar Structure */
 typedef struct FlowVar_ {
-    uint8_t type; /* type, DETECT_FLOWVAR in this case */
-    GenericVar *next; /* right now just implement this as a list,
-                       * in the long run we have think of something
-                       * faster. */
-    uint16_t idx; /* name idx */
+    uint8_t type;       /* type, DETECT_FLOWVAR in this case */
+    GenericVar *next;   /* right now just implement this as a list,
+                         * in the long run we have think of something
+                         * faster. */
+    uint16_t idx;       /* name idx */
     uint8_t datatype;
     union {
         FlowVarTypeStr fv_str;
@@ -59,12 +59,11 @@ typedef struct FlowVar_ {
 
 } FlowVar;
 
-
 /** Flowvar Interface API */
 
-void FlowVarAddStr(Flow *, uint8_t, uint8_t *, uint16_t);
-void FlowVarAddInt(Flow *, uint8_t, uint32_t);
-FlowVar *FlowVarGet(Flow *, uint8_t);
+void FlowVarAddStr(Flow *, uint16_t, uint8_t *, uint16_t);
+void FlowVarAddInt(Flow *, uint16_t, uint32_t);
+FlowVar *FlowVarGet(Flow *, uint16_t);
 void FlowVarFree(FlowVar *);
 void FlowVarPrint(GenericVar *);
